# crates/haneul-rpc-api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Haneul Node Api",
    "description": "REST Api for interacting with the Haneul Blockchain",
    "contact": {
      "name": "Haneul Labs",
      "url": "https://github.com/GeunhwaJeong/haneul"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
    },
    "version": "unknown"
  },
  "servers": [
    {
      "url": "/v2"
    }
  ],
  "paths": {
...
    "/checkpoints/{checkpoint}": {
      "get": {
        "tags": [
          "Checkpoint"
        ],
        "description": "[![stable](https://img.shields.io/badge/api-stable-53b576?style=for-the-badge)](#)\n\nFetch a Checkpoint\n\nFetch a checkpoint either by `CheckpointSequenceNumber` (checkpoint height) or by\n`CheckpointDigest` and optionally request its contents.\n\nIf the checkpoint has been pruned and is not available, a 410 will be returned.",
        "operationId": "Get Checkpoint",
        "parameters": [
          {
            "in": "path",
            "name": "checkpoint",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": ""
          },
          "410": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
    "/accounts/{account}/objects": {
      "get": {
        "tags": [
          "Account"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\n",
        "operationId": "ListAccountObjects",
        "parameters": [
          {
            "in": "path",
            "name": "account",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "x-haneul-cursor": {
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {}
            }
          }
        }
      }
    },
...
    "/checkpoints": {
      "get": {
        "tags": [
          "Checkpoint"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\nList Checkpoints\n\nRequest a page of checkpoints, and optionally their contents, ordered by\n`CheckpointSequenceNumber`.\n\nIf the requested page is below the Node's `lowest_available_checkpoint`, a 410 will be\nreturned.",
        "operationId": "List Checkpoints",
        "responses": {
          "200": {
            "description": "",
            "headers": {
              "x-haneul-cursor": {
                "style": "simple",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {}
            }
          },
          "410": {
            "description": ""
          },
          "500": {
            "description": ""
          }
        }
      }
    },
...
    "/coins/{coin_type}": {
      "get": {
        "tags": [
          "Coins"
        ],
        "description": "[![unstable](https://img.shields.io/badge/api-unstable-red?style=for-the-badge)](#) _Api subject to change; use at your own risk_\n\n",
        "operationId": "GetCoinInfo",
        "parameters": [
          {
            "in": "path",
            "name": "coin_type",
            "required": true,
            "schema": {
              "type": "string"
            },
            "style": "simple"
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {}
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
...
  },
...
}